Added clang-tidy configuration to the build
- Cleaned lint warnings from BaseBitFlag, BaseIterator, and Replicate
- Added a new --lint option to the build to make linting optional
- Added a new build flag DS_BUILD_LINT
- Changed type for constant::BYTESIZE to short int
- Changed properties to use training return types

+++ src/ds/constants.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 namespace constants {
-const unsigned int BYTESIZE = 8;
+const unsigned short int BYTESIZE = 8;
 const unsigned int BUFSIZE = 1024;
 const unsigned int READBUFSIZE = 8 * BUFSIZE;
 const char SEPARATOR = '/';
